# docs/swagger.yaml
basePath: /
definitions:
  internal_handlers.SwaggerBackupRequest:
    description: Request body for creating manual backups
    properties:
      backup_name:
        description: Optional backup name
        example: manual-backup-2024
        type: string
      immediate:
        description: Whether to create backup immediately
        example: true
        type: boolean
    type: object
  internal_handlers.SwaggerBackupResponse:
    description: Response for backup creation with details
    properties:
      data:
        properties:
          backup_path:
            example: /path/to/backup
            type: string
          backup_size_bytes:
            example: 1024000
            type: integer
          created_at:
            example: "2024-01-01T00:00:00Z"
            type: string
          message:
            example: Backup created successfully
            type: string
          success:
            example: true
            type: boolean
        type: object
      message:
        example: Backup created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  internal_handlers.SwaggerCleanupResponse:
    description: Response for access tracker cleanup operations
    properties:
      data:
        description: Cleanup operation data
        properties:
          duration_ms:
            description: Duration of the cleanup operation in milliseconds
            example: 25
            type: integer
          removed_trackers:
            description: Number of removed trackers
            example: 15
            type: integer
        type: object
      message:
        description: Response message
        example: Access tracker cleanup completed successfully
        type: string
      success:
        description: Operation success status
        example: true
        type: boolean
    type: object
...
  internal_handlers.SwaggerErrorResponse:
    description: Standard error response format
    properties:
      data:
        description: Error details
        properties:
          error:
            example: Content size exceeds maximum allowed size
            type: string
        type: object
      message:
        description: Error message
        example: Validation failed
        type: string
      success:
        description: Success indicator (always false for errors)
        example: false
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: |
    A high-performance content storage server with enterprise features including expiration, access limits, backup management, and sequential write processing.

    ## Configuration Impact on API Behavior

    The server behavior is heavily influenced by environment configuration:

    ### Authentication (ENABLE_AUTH)
    - When `ENABLE_AUTH=true`: All API endpoints except `/health`, `/health/detailed`, `/ping`, and `/debug/*` require authentication
    - Authentication methods: `X-API-Key` header or `api_key` query parameter
    - When `ENABLE_AUTH=false`: All endpoints are publicly accessible

    ### Content Size Limits (MAX_CONTENT_SIZE)
    - Default: 10MB (10485760 bytes)
    - All content storage operations validate against this limit
    - Exceeding this limit returns HTTP 400 with validation error

    ### Rate Limiting (THROTTLE_LIMIT, THROTTLE_BACKLOG_LIMIT)
    - `THROTTLE_LIMIT`: Maximum concurrent requests (default: 1000)
    - `THROTTLE_BACKLOG_LIMIT`: Maximum queued requests (default: 50)
    - Exceeding limits returns HTTP 429 Too Many Requests

    ### Content Type Restrictions (ALLOWED_CONTENT_TYPES)
    - Default: `application/json,text/plain`
    - Only specified content types are accepted for requests
    - Invalid content types return HTTP 415 Unsupported Media Type

...

...

...

    ### Write Queue System
...
paths:
  /api/v1/backup:
    post:
      consumes:
      - application/json
      description: |-
        Create a manual backup of the database with optional configuration, comprehensive logging, and detailed response information

        ## Configuration Dependencies:
        - **ENABLE_AUTH**: Requires authentication when enabled
        - **THROTTLE_LIMIT**: May return 429 if rate limit exceeded
        - **BACKUP_DIR**: Determines where backups are stored (from server configuration)
        - **Storage backend dependent** - Only supported for BadgerDB storage

        ## Request Parameters:
        - **backup_name**: Optional custom backup name (defaults to timestamp-based naming)
        - **immediate**: Optional flag to force immediate backup (defaults to true)
        - Invalid JSON in request body defaults to immediate backup with auto-generated name

        ## Backup Process:
        - Creates immediate backup regardless of BACKUP_INTERVAL schedule
        - Backup includes all current data, metadata, and access tracking information
        - Returns detailed backup information including path, size, and creation time
        - Operation duration varies with database size (logged for monitoring)
        - Comprehensive request logging with client IP, request ID, and user agent

        ## Response Information:
        - Backup path and file size from storage backend
        - Creation timestamp and operation duration
        - Backup statistics for monitoring and verification
      parameters:
      - description: Backup configuration (optional)
        in: body
        name: backup
        schema:
          $ref: '#/definitions/internal_handlers.SwaggerBackupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Backup created successfully
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerBackupResponse'
        "401":
          description: Authentication required (when ENABLE_AUTH=true)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded (THROTTLE_LIMIT)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "500":
          description: Backup operation failed
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "501":
          description: Backup not supported for storage type
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      summary: Create manual database backup with comprehensive monitoring
      tags:
      - Management Operations
  /api/v1/cleanup:
    post:
      consumes:
      - application/json
      description: |-
        Manually trigger cleanup of stale access trackers to prevent memory leaks

        ## Configuration Dependencies:
        - **ENABLE_AUTH**: Requires authentication when enabled
        - **THROTTLE_LIMIT**: May return 429 if rate limit exceeded
        - **Storage backend dependent** - Only supported for BadgerDB storage

        ## Behavior:
        - Removes stale access tracking data for non-existent content
        - Prevents memory leaks from accumulating access counters
        - Returns count of removed trackers and operation duration
        - Safe to run during normal operations
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup completed successfully
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerCleanupResponse'
        "401":
          description: Authentication required (when ENABLE_AUTH=true)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded (THROTTLE_LIMIT)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      summary: Cleanup access trackers
      tags:
      - Management Operations
...
  /api/v1/gc:
    post:
      consumes:
      - application/json
      description: |-
        Manually trigger database garbage collection to reclaim disk space and optimize performance

        ## Configuration Dependencies:
        - **ENABLE_AUTH**: Requires authentication when enabled
        - **THROTTLE_LIMIT**: May return 429 if rate limit exceeded
        - **Storage backend dependent** - Only supported for BadgerDB storage

        ## Behavior:
        - Forces immediate garbage collection on the database
        - Reclaims disk space from deleted or expired content
        - May temporarily impact performance during operation
        - Operation duration varies based on database size
      produces:
      - application/json
      responses:
        "200":
          description: Garbage collection completed successfully
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerGCResponse'
        "401":
          description: Authentication required (when ENABLE_AUTH=true)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded (THROTTLE_LIMIT)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "500":
          description: Garbage collection failed
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      summary: Trigger manual garbage collection
      tags:
      - Management Operations
...
  /api/v1/sync:
    post:
      consumes:
      - application/json
      description: |-
        Manually trigger a database synchronization operation with comprehensive logging and monitoring

        ## Configuration Dependencies:
        - **ENABLE_AUTH**: Requires authentication when enabled
        - **THROTTLE_LIMIT**: May return 429 if rate limit exceeded
        - **Storage backend dependent** - Sync behavior varies by storage implementation

        ## Current Implementation:
        - **Stub implementation**: No actual synchronization logic is performed
        - Comprehensive request logging with client IP, request ID, and user agent
        - Duration tracking for monitoring and performance analysis
        - Always returns success for compatibility with monitoring systems

        ## Future Enhancement:
        - Intended for database synchronization operations
        - May improve performance after heavy write operations
        - Could include write queue flushing and database consistency checks
      produces:
      - application/json
      responses:
        "200":
          description: Sync completed successfully
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerSuccessResponse'
        "401":
          description: Authentication required (when ENABLE_AUTH=true)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded (THROTTLE_LIMIT)
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
        "500":
          description: Sync operation failed
          schema:
            $ref: '#/definitions/internal_handlers.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      - ApiKeyQuery: []
      summary: Trigger manual database synchronization (stub implementation)
      tags:
      - Management Operations
...
schemes:
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication. Required when ENABLE_AUTH=true. Can also
      be provided as 'api_key' query parameter.
    in: header
    name: X-API-Key
    type: apiKey
  ApiKeyQuery:
    description: API key authentication via query parameter. Alternative to X-API-Key
      header.
    in: query
    name: api_key
    type: apiKey
swagger: "2.0"